Cleaned up gauge series.
