+++ ts/Series/Gauge/GaugePointOptions.d.ts
@@ -27,7 +27,7 @@ interface GaugePointOptions extends LinePointOptions {
 
 /* *
  *
- *  Default export
+ *  Default Export
  *
  * */
